Flytt api test config til egen klasse
